# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "此 API 文件提供社區通知系統的後端介面說明，包含使用者管理、設施預約、社區申請及訊息通知等功能。使用 JWT Bearer 進行認證。",
        "title": "社區通知系統 API 文件 (Community Notification System)",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/community/getlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "依縣市、行政區、郵遞區號或關鍵字等條件查詢社區基本資料清單，並支援分頁查詢。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CommunityManager"
                ],
                "summary": "取得社區列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "縣市過濾（例如：新北市）",
                        "name": "municipality",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "行政區過濾（例如：淡水區）",
                        "name": "district",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "郵遞區號過濾（例如：251）",
                        "name": "postal_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "關鍵字搜尋（社區名稱或地址）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "頁碼（從 1 開始，預設 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數（預設 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功取得社區清單",
                        "schema": {
                            "$ref": "#/definitions/community.CommunityListResponse"
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/community/register": {
            "post": {
                "description": "社區申請人填寫基本資料送出申請，狀態預設為 pending。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CommunityManager"
                ],
                "summary": "社區送出申請",
                "parameters": [
                    {
                        "description": "社區申請資料",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/community.RegisterApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "社區申請已送出",
                        "schema": {
                            "$ref": "#/definitions/community.RegisterApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/community/register/{id}/approve": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "僅允許 Super admin 操作，核可後建立正式社區與初始 admin 帳號。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CommunityManager"
                ],
                "summary": "核可社區申請",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "核可成功，回傳社區與管理員資訊",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "請求錯誤或無此申請單",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "403": {
                        "description": "權限不足",
                        "schema": {
                            "$ref": "#/definitions/model.Response403Error"
                        }
                    },
                    "500": {
                        "description": "核可流程失敗",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/community/register/{id}/reject": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "僅允許 Super admin 操作，駁回申請並提供原因。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CommunityManager"
                ],
                "summary": "駁回社區申請",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "駁回原因",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/community.RejectApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已駁回申請",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "請求錯誤或無此申請單",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "403": {
                        "description": "權限不足",
                        "schema": {
                            "$ref": "#/definitions/model.Response403Error"
                        }
                    },
                    "500": {
                        "description": "駁回流程失敗",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "community.CommunityListResponse": {
            "type": "object",
            "properties": {
                "communities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/community.CommunitySummary"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "community.CommunitySummary": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "251新北市淡水區濱海路一段306巷"
                },
                "alley_number": {
                    "type": "integer",
                    "example": 0
                },
                "community_id": {
                    "type": "integer",
                    "example": 1
                },
                "community_name": {
                    "type": "string",
                    "example": "甜水郡社區"
                },
                "district": {
                    "type": "string",
                    "example": "淡水區"
                },
                "lane_number": {
                    "type": "integer",
                    "example": 306
                },
                "municipality": {
                    "type": "string",
                    "example": "新北市"
                },
                "postal_code": {
                    "type": "integer",
                    "example": 251
                },
                "road_name": {
                    "type": "string",
                    "example": "濱海路一段"
                }
            }
        },
        "community.RegisterApplicationRequest": {
            "type": "object",
            "required": [
                "address",
                "admin_email",
                "admin_name",
                "admin_password",
                "applicant_email",
                "applicant_name",
                "community_name",
                "district",
                "municipality",
                "postal_code",
                "road_name"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "251新北市淡水區濱海路一段306巷"
                },
                "admin_email": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "admin_name": {
                    "type": "string",
                    "example": "AdminName"
                },
                "admin_password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                },
                "admin_phone": {
                    "type": "string",
                    "example": "0912345678"
                },
                "alley_number": {
                    "type": "integer",
                    "example": 0
                },
                "applicant_email": {
                    "type": "string",
                    "example": "applicant@example.com"
                },
                "applicant_name": {
                    "type": "string",
                    "example": "ApplicantName"
                },
                "applicant_phone": {
                    "type": "string",
                    "example": "0987654321"
                },
                "community_name": {
                    "type": "string",
                    "example": "甜水郡社區"
                },
                "district": {
                    "type": "string",
                    "example": "淡水區"
                },
                "lane_number": {
                    "type": "integer",
                    "example": 306
                },
                "municipality": {
                    "type": "string",
                    "example": "新北市"
                },
                "postal_code": {
                    "type": "integer",
                    "example": 251
                },
                "remark": {
                    "type": "string",
                    "example": "申請建立測試社區"
                },
                "road_name": {
                    "type": "string",
                    "example": "濱海路一段"
                }
            }
        },
        "community.RegisterApplicationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "application_id": {
                            "type": "integer",
                            "example": 1
                        },
                        "status": {
                            "type": "string",
                            "example": "pending"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "社區申請已送出"
                }
            }
        },
        "community.RejectApplicationRequest": {
            "type": "object",
            "properties": {
                "reject_reason": {
                    "type": "string",
                    "example": "資訊不全，請重新申請"
                }
            }
        },
...
        "model.Response400Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid input data"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1001
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "model.Response401Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "error": {
                    "type": "string",
                    "example": "Invalid or missing JWT token"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1002
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
        "model.Response403Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "error": {
                    "type": "string",
                    "example": "Permission denied"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1003
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Forbidden"
                }
            }
        },
...
        "model.Response500Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1000
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}